Separate Packet and analyze_packet
